requirements.txt
PyYAML==6.0.1
GitPython==3.1.40
requests==2.31.0
python-dotenv==1.0.0
watchdog==3.0.0
schedule==1.2.0
colorlog==6.7.0
